pyproject.toml
[project]
name = "video-upload-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "pillow>=12.0.0",
    "python-decouple>=3.8",
    "requests>=2.32.5",
]
